Address decompilation review across HUD, collisions, and screen transitions
Clarify types, constants, private data, and transition semantics. Retain original Kirby contact assembly behind NONMATCHING and provide typed C for the remaining instruction mismatch.

AI-assisted implementation and review coordinated with three Codex subagents.

+++ include/main.h
@@ -47,8 +47,8 @@ struct BlendRegs {
 
 struct Unk_03002EC0 {
     /* TODO: fixing types breaks sub_081525DC. */
-    u32 unk0; // const void *
-    u32 unk4; // void *
+    uintptr_t unk0; // const void *
+    uintptr_t unk4; // void *
     u16 unk8;
     u8 fillerA[2];
 };
